+++ src/ObjectCache/Exception/CachedTargetObjectNotFoundException.php
@@ -17,4 +17,7 @@ use Rekalogika\Mapper\Exception\RuntimeException;
 
 class CachedTargetObjectNotFoundException extends RuntimeException
 {
+    public function __construct()
+    {
+    }
 }